+++ lib/it/link.rb
@@ -7,7 +7,7 @@ module It
 
     # See It.link for details. You can do everything there and save 6 characters.
     def initialize(href, options = {})
-      raise TypeError, 'Invalid href given' unless [Hash, String].include?(href.class)
+      raise TypeError, 'Invalid href given' unless [Hash, String, ActiveSupport::SafeBuffer].include?(href.class)
 
       super(:a, options)
       @href = href